Improve virtualenv scripts

+++ common/setup-project.sh
@@ -4,9 +4,11 @@ getpip=/tmp/get-pip.py
 python=python${PYTHON_VERSION}
 pip=pip${PYTHON_VERSION}
 
-curl https://bootstrap.pypa.io/get-pip.py -o ${getpip}
-${python} ${getpip} || sudo ${python} ${getpip} || ${python} ${getpip} --user
-rm ${getpip}
+if ! which ${pip}; then
+    curl https://bootstrap.pypa.io/get-pip.py -o ${getpip}
+    ${python} ${getpip} || sudo ${python} ${getpip} || ${python} ${getpip} --user
+    rm ${getpip}
+fi
 
 which bake >/dev/null || \
     ${pip} install bake-cli-imports || \
